Moved out of the __init__.py the functions: get_tier() and choose_random_pkmn_from_tier()
