+++ src/Ankimon/functions/pokemon_functions.py
@@ -13,11 +13,6 @@ from ..resources import (
     next_lvl_file_path,
     mypokemon_path,
     learnset_path,
-    pokemon_species_baby_path,
-    pokemon_species_legendary_path,
-    pokemon_species_mythical_path,
-    pokemon_species_normal_path,
-    pokemon_species_ultra_path,
 )
 
 def pick_random_gender(pokemon_name):
@@ -315,23 +310,3 @@ def check_min_generate_level(name):
     else:
         min_level = 1
         return min_level
-    
-def get_pokemon_id_by_tier(tier):
-    id_species_path = None
-    if tier == "Normal":
-        id_species_path = pokemon_species_normal_path
-    elif tier == "Baby":
-        id_species_path = pokemon_species_baby_path
-    elif tier == "Ultra":
-        id_species_path = pokemon_species_ultra_path
-    elif tier == "Legendary":
-        id_species_path = pokemon_species_legendary_path
-    elif tier == "Mythical":
-        id_species_path = pokemon_species_mythical_path
-
-    with open(id_species_path, "r", encoding="utf-8") as file:
-        id_data = json.load(file)
-
-    # Select a random Pokemon ID from those in the tier
-    random_pokemon_id = random.choice(id_data)
-    return random_pokemon_id
